Render the framebuffer using a texture
Instead of glDrawPixels. It's more efficient, since we can
cache the texture on the GPU, rather than uploading it every
frame. This allows us to have an interactive GUI, if we need to.

TODO: Figure out how often to upload the framebuffer. ATM, it's every
time it changes. But I feel like that will be at least every frame. In which
case, it's pointless to even test for it. That said, since this is only
meant as a visualizer, we could perhaps do a time-based update. And we
could force an update on camera changes, so camera movements are smooth.