Cache cartesian camera axes
A quick Very Sleepy profile run showed CalculateRayFromPixel()
to be taking the same amount of time per frame as scene intersection.
This should be very obvious, since we're doing a ton of math
over and over again, for each pixel. The original camera
code came from a real-time engine, where it is expected for the
camera to move quite often. However, for path tracing, the camera
should move very infrequently, if at all. After implementing
the caching, scene traversal is 88% of the frame time, and
CalculateRayFromPixel() is 0.11%