Use vectors for FrameBuffer data instead of raw arrays
When we call Scene::SetCamera, we use the default assignment operator to
replace the old camera with the new one. With raw arrays, this was just
copying over the pointers. This lead to memory access exceptions, since
the pointer is deleted by the destructor. Rather than go through the
hassle of creating my own assignment operators, etc. we can just use
vectors, which handle the move automagically.