Increase the maximum number of bounces
I used the 'furnace test' to test if the IdealSpecularDielectricBSDF
was working correctly. Essentially, I just set the albedo to 1.0, and
removed any explicit lights. In this setup, the object should be
invisible, since the surface albedo doesn't attenuate anything. However,
this wasn't the case. There was shadowing around some of the scales/edges
of the dragon.

Essentially, a ray would enter, and due to total internal reflection,
would never find a path for exit. By increasing the max path length, we
increase the number of chances a path has to find a way out.
The dragon model only needed about 25 max path length, but since we have
Russian Roulette set up, increasing to 40 doesn't really hurt us.