Reduce the default FOV to 45 degrees
I did some searching around about the FOVs of actual cameras, and they
widely vary depending mostly on the lens size, but also on the brand.

But a general spectrum is perhaps 25 - 90 degrees, with 90 degrees being
a large wide angle lens.
http://forums.cgsociety.org/archive/index.php?t-330980.html
"Common for architectural photography of a room would be a typical
wide-angle lens of about 18mm to 28mm (or a 90 degree to 65 degree FOV.)"

"Common for pictures of medium-sized subjects, for a catalog or
something, would be about a 50mm lens (or 40 degree FOV.)"

"Common for a portrait of a person would be about an 80mm lens (or
25 degree FOV.)"

"These are for stills on 35mm film."

Given this info and the current objects I was rendering, 45 degree FOV
seemed to be more reasonable. If, in the future, I need a different
FOV, it can be specified in the camera constructor.