Move all the "rendering" into a new integrator class
The integrator and visualizer use a set of FrameData structs to 
pass data back and forth. It's structured very similarily to Vulkan's
VK_PRESENT_MODE_MAILBOX_KHR. The integrator will keep
generating data, and between each frame, it will swap which buffer
it's using with an atomic pointer.

Then the visualizer, on it's frames, will swap its buffer with the atomic
pointer, and the accumulate all the data. And show that on the screen.


NOTE: In the code, it's currently set to *not* accumulate. Since it's
doing an animation, not Monte Carlo Rendering.