Remove "extension" lines
There no longer needed. And the validator will
give a warning now.

```
This extension is not enabled with a #extension as other extensions are.
It is also not enabled through use of a profile or #version.  The intended
level of GLSL/ESSL features, independent from Vulkan-specific usage, comes
from the traditional use of #version, profile, and #extension.


Instead, use of this extension is an effect of using a GLSL front-end in a
mode that has it generate SPIR-V for Vulkan.  Such tool use is outside the
scope of using the Vulkan API and outside the definition of GLSL and this
extension. See the documentation of the compiler to see how to request
generation of SPIR-V for Vulkan.
```