Convert integrator code to use ISPC
It was a bit of a PITA, since ISPC is c-based, with no classes. So we have to do more public structs with free functions.

But the payoff was great. The compiler still shows lots of "performance" warnings. But already the integrator went from 35ms per frame to 15ms per frame.