Switch to our own fork of IMGUI
Git submodules need to be able to clone from a public version of the commit. Aka, any changes we make in the submodule repo need to be pushed somewhere. So another client can reference the same commit.

We edit Dear IMGUI to add a CMakeLists.txt file. We fork to do it.